+++ src/main/java/insane96mcp/survivalreimagined/module/experience/enchantments/enchantment/Blasting.java
@@ -1,6 +1,8 @@
 package insane96mcp.survivalreimagined.module.experience.enchantments.enchantment;
 
 import insane96mcp.survivalreimagined.setup.SREnchantments;
+import net.minecraft.world.effect.MobEffectUtil;
+import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
@@ -8,7 +10,9 @@ import net.minecraft.world.item.PickaxeItem;
 import net.minecraft.world.item.enchantment.DiggingEnchantment;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
+import net.minecraft.world.item.enchantment.EnchantmentHelper;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraftforge.common.ForgeMod;
 
 public class Blasting extends Enchantment {
     static final EnchantmentCategory PICKAXES = EnchantmentCategory.create("pickaxes", item -> item instanceof PickaxeItem);
@@ -44,6 +48,30 @@ public class Blasting extends Enchantment {
         if (level == 0)
             return 0f;
 
-        return (float) (level * Math.pow(2.5d, 6.5f - state.getBlock().getExplosionResistance()));
+        float miningSpeedBoost = (float) (level * Math.pow(2.5d, 6.5f - state.getBlock().getExplosionResistance()));
+
+        if (MobEffectUtil.hasDigSpeed(entity)) {
+            miningSpeedBoost *= 1.0F + (float)(MobEffectUtil.getDigSpeedAmplification(entity) + 1) * 0.2F;
+        }
+
+        if (entity.hasEffect(MobEffects.DIG_SLOWDOWN)) {
+            float miningFatigueMultiplier = switch (entity.getEffect(MobEffects.DIG_SLOWDOWN).getAmplifier()) {
+				case 0 -> 0.3F;
+				case 1 -> 0.09F;
+				case 2 -> 0.0027F;
+				default -> 8.1E-4F;
+			};
+			miningSpeedBoost *= miningFatigueMultiplier;
+        }
+
+        if (entity.isEyeInFluidType(ForgeMod.WATER_TYPE.get()) && !EnchantmentHelper.hasAquaAffinity(entity)) {
+            miningSpeedBoost /= 5.0F;
+        }
+
+        if (!entity.onGround()) {
+            miningSpeedBoost /= 5.0F;
+        }
+
+        return miningSpeedBoost;
     }
 }